+++ app/Http/Resources/Show.php
@@ -18,6 +18,8 @@ class Show extends JsonResource
             'id' => $this->id,
             'title' => $this->title,
             'description' => $this->description,
+            'track' => $this->track->name,
+            'term' => $this->term_id,
             'djs' => $this->hosts->pluck('full_name'),
             'start' => $this->start,
             'end' => $this->end,